refactor(core)!: 🔩 Change package name and look and feel
BREAKING CHANGE: New name and new command

- The package name was changed to Git Cook ✨
- The new command is Usage: gcook [options] [command]

CLI to manage git actions easily

Options:
  -V, --version   output the version number
  -h, --help      display help for command

Commands:
  check           Checks if you are up to date
  stats           Check how many downloads has the tool
  init [options]  Wizard to create a conventional commit
  help [command]  display help for command
- The look and feel was improved
