+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@vennet/gitool",
+  "name": "@vennet/gitcook",
   "version": "0.4.5",
-  "description": "",
+  "description": "CLI tool for git actions",
   "type": "module",
   "main": "./dist/index.js",
   "keywords": [
@@ -14,14 +14,14 @@
   "license": "ISC",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/vennet-developers/gitool.git"
+    "url": "git+https://github.com/vennet-developers/gitcook.git"
   },
   "bugs": {
-    "url": "https://github.com/vennet-developers/gitool/issues"
+    "url": "https://github.com/vennet-developers/gitcook/issues"
   },
-  "homepage": "https://github.com/vennet-developers/gitool#readme",
+  "homepage": "https://github.com/vennet-developers/gitcook#readme",
   "bin": {
-    "gt": "./dist/index.js"
+    "gcook": "./dist/index.js"
   },
   "scripts": {
     "dev": "rm -rf dist && tsc -w",
@@ -29,8 +29,8 @@
     "lint": "biome lint --apply src/*",
     "lint:check": "biome check --apply src/*",
     "format": "biome format --write src/*",
-    "pkg:link": "(pnpm --global unlink @vennet/gitool || true) && chmod +x ./dist/index.js && pnpm --global link",
-    "pkg:unlink": "pnpm --global unlink @vennet/gitool",
+    "pkg:link": "(pnpm --global unlink @vennet/gitcook || true) && chmod +x ./dist/index.js && pnpm --global link",
+    "pkg:unlink": "pnpm --global unlink @vennet/gitcook",
     "pkg:version": "changeset && changeset version && git add -A && git commit -am 'New bump version generated'",
     "pkg:publish": "changeset publish && git push && git push --tags"
   },
